+++ packages/vuikit/CHANGELOG.md
@@ -5,6 +5,7 @@
  - Fixed `vk-navbar-full` transparent modifier.
  - Fixed and changed `v-position` constants export.
  - Improved `vk-pagination` responsivnes by applying `vk-margin` directive.
+ - Fixed env issue affecting ES and CJS builds
 
 ## 0.8.2